# [WIP] cythonize plum.parametric