--- a/setup.py
+++ b/setup.py
@@ -6,7 +6,9 @@ setup(
     packages=find_packages(exclude=["docs"]),
     python_requires=">=3.6",
     install_requires=requirements,
-    ext_modules=[Extension("plum.function", ["plum/function.py"]), 
-                 Extension("plum.parametric", ["plum/parametric.py"])],
+    ext_modules=[
+        Extension("plum.function", ["plum/function.py"]),
+        Extension("plum.parametric", ["plum/parametric.py"]),
+    ],
     include_package_data=True,
 )