I experimented with cythonizing the plum.parametric module to speed it up a bit (for my own personal peace of mind).

Other than slightly modifying the test because now `typeof` is a function instead of a `function`, there is a change needed:

In Cython, type hints are all strings. So the `type_of` dispatches of `typeof(obj: list)` and `typeof(obj: tuple)` in `plum.parametric` will have a string and type instead of a real class, thus in type.py they will be forward references/promises which are never _fullfilled_.

Apparently plum breaks down when those two promises are not fulfilled.
A fix was to hardcode those two cases in ptype:
```python
def ptype(obj):

    if hasattr(obj, "__module__") and obj.__module__ == "typing":
        ...
    if isinstance(obj, str):
        if obj == "list":
            return ptype(list)
        elif obj == "tuple":
            return ptype(tuple)
        else:
            return get_forward_reference(obj)
```

there's probably a better way to address this, but I can't see it now.
The PromisedType mechanism is a bit confusing...