@PhilipVinc 2021-07-06
@wesselb do you have benchmarks i can try ?