@wesselb 2021-07-07
Another way of addressing the type-hints-become-strings issue is to use `dispatch.multi`, which doesn't use the type hints:
```
--- a/plum/parametric.py
+++ b/plum/parametric.py
@@ -280,12 +280,12 @@ def type_of(obj):
     return ptype(type(obj))


-@_dispatch
+@_dispatch.multi((list,))
 def type_of(obj: list):
     return List(_types_of_iterable(obj))


-@_dispatch
+@_dispatch.multi((tuple,))
 def type_of(obj: tuple):
     return Tuple(*(type_of(x) for x in obj))
```

I agree that the promise system is a confusing, but unfortunately it's necessary because there are some circular dependencies between the files.

There is a [small benchmark](https://github.com/wesselb/plum/blob/master/benchmark.py) in the root of the repo. Unfortunately, it looks like Cythonising `parametric.py` does not make much of a difference.

Without Cythonising:

```
# Function Calls
Native call:       0.08 us (1.0 x)
First plum call: 122.31 us (1538.0 x)
Plum call:         0.33 us (4.0 x)

# Parametric Function Calls
Native call:       0.07 us (1.0 x)
First plum call: 450.13 us (6565.0 x)
Plum call:        10.80 us (157.0 x)

# Class Calls
Native call:       0.09 us (1.0 x)
First plum call: 114.20 us (1269.0 x)
Plum call:         1.05 us (12.0 x)

# Class Attribute Calls
Native call:       0.14 us (1.0 x)
First plum call: 165.22 us (1178.0 x)
Plum call:         1.14 us (8.0 x)
```

With Cythonising:
```
# Function Calls
Native call:       0.09 us (1.0 x)
First plum call: 132.80 us (1517.0 x)
Plum call:         0.32 us (4.0 x)

# Parametric Function Calls
Native call:       0.07 us (1.0 x)
First plum call: 474.93 us (6593.0 x)
Plum call:         8.98 us (125.0 x)

# Class Calls
Native call:       0.09 us (1.0 x)
First plum call: 112.77 us (1230.0 x)
Plum call:         1.06 us (12.0 x)

# Class Attribute Calls
Native call:       0.14 us (1.0 x)
First plum call: 114.92 us (826.0 x)
Plum call:         1.15 us (8.0 x)
```

The big performance bottleneck is `type_of`. Currently, [there is a performance optimisation in place](https://github.com/wesselb/plum/blob/master/plum/function.py#L452) which only calls `type_of` if parametric types are being used. But this is too cautious: `type` will work just fine with `@parametric` classes. I can see a few scenarios where a mechanism like `type_of` cannot be avoided:

* parametric lists (`List[int]`), because `type([1, 1])` will always give just `list`;
* parametric tuples (`Tuple[int, int]`), for similar reasons; and
* constructions [like this](https://github.com/wesselb/plum/#hooking-into-type-inference).

What we probably need is a better algorithm to determine if `type_of` if necessary, not just use it whenever a parametric types are being used. What are your thoughts?