@wesselb 2023-03-03
Closing this since we've decided to go for a pure-Python package without Cythonisation.